+++ CHANGELOG.md
@@ -2,10 +2,16 @@
 
 ## Unreleased
 
-## 2.14.0
+## 3.0.0
 
 ### Added
 
+- Added support for changing the default Python recursion depth
+  ([#312](https://github.com/fortran-lang/fortls/issues/312))
+- Added support for preprocessor macro expansions
+  ([#368](https://github.com/fortran-lang/fortls/pull/368))
+- Added support for leading white spaces in preprocessor directives
+  ([#297](https://github.com/fortran-lang/fortls/issues/297))
 - Added hover messages for Types and Modules
   ([#208](https://github.com/fortran-lang/fortls/issues/208))
 - Added support for Markdown intrinsics from the M_intrinsics repository